Prefer transient popup to change key expiry date
This is a trade-off.

When the calendar of the WDateEdit is clicked, widget PopupExpiryTime gets hidden, forcing to show it again.

If PopupExpiryTime::setTransient is false, it will obviously persist, until explicitly hidden with the Close button. Not satisfactory to change key selection while it is visible, nor to do other tasks like changing certification trust level, certifying uids.... whatever.

Forcing its persistence while playing with many events has not been successful, and quite ugly.